@@ -6,17 +6,17 @@ namespace Test {
 
     public class TestConnectableProperty {
 
-        private readonly StoredProperty<int> parent = new StoredProperty<int>(1);
+        private readonly StoredProperty<int> parent = new(1);
 
         [Fact]
         public void InitialValue() {
-            var connectableProperty = new ConnectableProperty<int>(0);
+            var connectableProperty = new ConnectableProperty<int>();
             connectableProperty.Value.Should().Be(0, "before connecting");
         }
 
         [Fact]
         public void ConnectingChangesValue() {
-            var connectableProperty = new ConnectableProperty<int>(0);
+            var connectableProperty = new ConnectableProperty<int>();
 
             connectableProperty.Connect(parent);
             connectableProperty.Value.Should().Be(1);
@@ -24,8 +24,8 @@ namespace Test {
 
         [Fact]
         public void ConnectingWithDifferentValueFiresChange() {
-            bool eventFired = false;
-            var connectableProperty = new ConnectableProperty<int>(0);
+            bool eventFired          = false;
+            var  connectableProperty = new ConnectableProperty<int>();
             connectableProperty.PropertyChanged += delegate { eventFired = true; };
 
             connectableProperty.Connect(parent);
@@ -34,17 +34,17 @@ namespace Test {
 
         [Fact]
         public void ConnectingWithSameValueDoesNotFireChange() {
-            bool eventFired = false;
-            var connectableProperty = new ConnectableProperty<int>(0);
+            bool eventFired          = false;
+            var  connectableProperty = new ConnectableProperty<int>();
             connectableProperty.PropertyChanged += delegate { eventFired = true; };
 
-            connectableProperty.Connect(new StoredProperty<int>(0));
+            connectableProperty.Connect(new StoredProperty<int>());
             eventFired.Should().BeFalse();
         }
 
         [Fact]
         public void ParentValueChangesTriggerValueChangeInConnectableProperty() {
-            var connectableProperty = new ConnectableProperty<int>(0);
+            var connectableProperty = new ConnectableProperty<int>();
             connectableProperty.Connect(parent);
 
             parent.Value = 2;
@@ -53,8 +53,8 @@ namespace Test {
 
         [Fact]
         public void ParentValueChangesTriggerChangeEventsInConnectableProperty() {
-            bool eventFired = false;
-            var connectableProperty = new ConnectableProperty<int>(0);
+            bool eventFired          = false;
+            var  connectableProperty = new ConnectableProperty<int>();
             connectableProperty.Connect(parent);
             connectableProperty.PropertyChanged += delegate { eventFired = true; };
             eventFired.Should().BeFalse("haven't changed the parent property yet");
@@ -65,7 +65,7 @@ namespace Test {
 
         [Fact]
         public void DisconnectingResetsValueToDisconnectedValue() {
-            var connectableProperty = new ConnectableProperty<int>(0);
+            var connectableProperty = new ConnectableProperty<int>();
             connectableProperty.Connect(parent);
             connectableProperty.Value.Should().Be(1, "before disconnecting");
 
@@ -75,8 +75,8 @@ namespace Test {
 
         [Fact]
         public void DisconnectingFiresChange() {
-            bool eventFired = false;
-            var connectableProperty = new ConnectableProperty<int>(0);
+            bool eventFired          = false;
+            var  connectableProperty = new ConnectableProperty<int>();
             connectableProperty.Connect(parent);
             connectableProperty.PropertyChanged += delegate { eventFired = true; };
             eventFired.Should().BeFalse("have not disconnected yet");
@@ -87,8 +87,8 @@ namespace Test {
 
         [Fact]
         public void DisconnectingWithSameValueResetsValueToDisconnectedValue() {
-            var connectableProperty = new ConnectableProperty<int>(0);
-            connectableProperty.Connect(new StoredProperty<int>(0));
+            var connectableProperty = new ConnectableProperty<int>();
+            connectableProperty.Connect(new StoredProperty<int>());
             connectableProperty.Value.Should().Be(0, "before disconnecting");
 
             connectableProperty.Disconnect();
@@ -97,14 +97,32 @@ namespace Test {
 
         [Fact]
         public void DisconnectingWithSameValueDoesNotFireChange() {
-            bool eventFired = false;
-            var connectableProperty = new ConnectableProperty<int>(0);
-            connectableProperty.Connect(new StoredProperty<int>(0));
+            bool eventFired          = false;
+            var  connectableProperty = new ConnectableProperty<int>();
+            connectableProperty.Connect(new StoredProperty<int>());
             connectableProperty.PropertyChanged += delegate { eventFired = true; };
             eventFired.Should().BeFalse("have not disconnected yet");
 
             connectableProperty.Disconnect();
             eventFired.Should().BeFalse("disconnected, but the value should not have changed from 0 to 0");
+        }
+
+        [Fact]
+        public void ConnectingConstantValueChangesValue() {
+            var connectableProperty = new ConnectableProperty<int>();
+
+            connectableProperty.Connect(2);
+            connectableProperty.Value.Should().Be(2);
+        }
+
+        [Fact]
+        public void ConnectConstantValueFiresChange() {
+            bool eventFired          = false;
+            var  connectableProperty = new ConnectableProperty<int>();
+            connectableProperty.PropertyChanged += delegate { eventFired = true; };
+
+            connectableProperty.Connect(2);
+            eventFired.Should().BeTrue();
         }
 
     }
