@@ -1,4 +1,6 @@
-﻿using System;
+﻿// ReSharper disable RedundantUsingDirective - conditional compilation sometimes makes Visual Studio mark the wrong code as unused, reloading project helps
+
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
@@ -14,13 +16,16 @@ namespace KoKo.Property {
 
         private static readonly IDictionary<Type[], Type> EventProxyStructCache = new Dictionary<Type[], Type>();
 
+#if !NETSTANDARD2_0
         private static ModuleBuilder? _moduleBuilder;
         private static long           _classNameCounter;
-
         public event EventHandler? OnEvent;
+#endif
 
         /// <exception cref="ArgumentException">If the named event cannot be found on the given object.</exception>
+        /// <exception cref="PlatformNotSupportedException">If you are running on .NET Standard 2.0. To use this class, target at least .NET Framework 4.5.2, .NET Standard 2.1, or .NET Core 3.0.</exception>
         public NativeEventListener(object nativeObject, string nativeEventName) {
+#if !NETSTANDARD2_0
             EventInfo? nativeEvent = nativeObject.GetType().GetTypeInfo().GetEvent(nativeEventName);
 
             Type?       handlerType  = nativeEvent?.EventHandlerType;
@@ -62,6 +67,10 @@ namespace KoKo.Property {
 
             Delegate eventHandlerDelegate = Delegate.CreateDelegate(handlerType, eventProxyStruct, eventReceiverMethod);
             nativeEvent.AddEventHandler(nativeObject, eventHandlerDelegate);
+# else
+            throw new PlatformNotSupportedException("NativeEventListener is not supported on .NET Standard 2.0 because ModuleBuilder is unavailable. To use this class, please target .NET Framework " +
+                "4.5.2 or later, or .NET Standard 2.1, or .NET (Core) 3.0 or later.");
+#endif
         }
 
     }
