@@ -6,7 +6,7 @@ namespace KoKo.Property {
 
     /// <summary>A property that can interoperate with native read-write C# properties that use <c>INotifyPropertyChanged</c>. For readonly C# properties, see <see cref="NativeReadableProperty{T}" />.<br />If you change this object's Value, that change will be propagated to the backing object.</summary>
     /// <typeparam name="T">The type of this property's value, which is also the type of the native C# property</typeparam>
-    public class NativeWritableProperty<T>: SettableProperty<T> {
+    public class NativeWritableProperty<T>: AbstractSettableProperty<T> {
 
         private readonly object       cachedValueLock = new();
         private readonly object       nativeObject;
@@ -76,10 +76,15 @@ namespace KoKo.Property {
         /// <param name="nativePropertyName">The name of a regular C# property (not a KoKo property) on the <c>nativeObject</c> that triggers an event on the object.<br />To be more type-safe here, you can use <c>nameof(MyNativeObjectClass.MyNativeProperty)</c> instead of a string <c>"MyNativeProperty"</c>.</param>
         /// <param name="nativeEventName">The name of the event that is raised on <c>nativeObject</c> when the value of the <c>nativePropertyName</c> property is changed.<br />To be more type-safe here, you can use <c>nameof(MyNativeObjectClass.MyNativePropertyChanged)</c> instead of a string <c>"MyNativePropertyChanged"</c>.<br />If this parameter is omitted or null, it defaults to appending <c>"Changed"</c> to the <c>nativePropertyName</c> parameter, e.g. <c>new NativeWritableProperty&lt;string&gt;(myToolStripItem, nameof(ToolStripItem.Text))</c> will listen for <c>TextChanged</c> events on <c>myToolStripItem</c>.</param>
         /// <exception cref="ArgumentException">if the native property does not exist on the given object, or if its getter isn't accessible, or if its value's type does not match <code>&lt;T&gt;</code></exception>
+        /// <exception cref="PlatformNotSupportedException">On .NET Standard 2.0; try the other constructor that takes an <see cref="INotifyPropertyChanged"/> argument, or use at least .NET Framework 4.5.2, .NET Standard 2.1, or .NET (Core) 3.1.</exception>
         public NativeWritableProperty(object nativeObject, string nativePropertyName, string? nativeEventName = null): this(nativeObject, nativePropertyName) {
+#if !NETSTANDARD2_0
             nativeEventName ??= nativePropertyName + "Changed";
             NativeEventListener nativeEventListener = new(nativeObject, nativeEventName);
             nativeEventListener.OnEvent += delegate { NativePropertyChanged(); };
+#else
+            throw new PlatformNotSupportedException("Constructor NativeWritableProperty(object nativeObject, string nativePropertyName, string? nativeEventName = null) is not supported on .NET Standard 2.0 because ModuleBuilder is unavailable. If your native property implements INotifyPropertyChanged, you can use the alternate constructor NativeWritableProperty(INotifyPropertyChanged nativeObject, string nativePropertyName). Alternately you can target .NET Framework 4.5.2 or later, or .NET Standard 2.1, or .NET (Core) 3.0 or later.");
+#endif
         }
 
         private void NativePropertyChanged(object? sender = null, PropertyChangedEventArgs? e = null) {
